requirements.txt
torch>=2.7.0
torchvision>=0.22.0
PyYAML>=6.0.1
tqdm>=4.67.1
Pillow>=10.4.0
numpy>=2.2.0
